Fix bad values in Gloat.UnappliedMigrations
Say you work in a branch called `feature` where you add a new migration.
You have the following migrations applied in the branch feature.

Applied: [1, 2, 3, 4] | Available: [1, 2, 3, 4]

Migration 4 is only available in the branch `feature.

Now, if you go back to a branch where we don't have the newly applied
migration, say `master` you have.

Applied: [1, 2, 3, 4] | Available: [1, 2, 3]

Previously, gloat would have tried to apply the migration with version 4
on master, which is problematic, because it is already applied and not
available again for application. The workflow I have showed above is
pretty common thing to do, and now gloat respects that.